Paper-Scrapper/requirements.txt
arxiv>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
